:truck: refactor input folder structure #42
