+++ src/components/input/types.d.ts
@@ -1,5 +1,3 @@
 import * as constants from 'constants';
 
 export type inputSizeType = (typeof constants.INPUT_SIZE_TYPE)[number];
-
-export type inputType = (typeof constants.INPUT_TYPE)[number];
